backend/requirements.txt
fastapi
uvicorn
supabase
python-multipart
httpx
exponent-server-sdk
openai
slowapi
requests
psycopg2-binary
python-dotenv
